// SDKs/twilio.com/twilio_trunking_v1/1.36.0/typescript/src/sdk/models/shared/trunkingv1trunkipaccesscontrollist.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";


export class TrunkingV1TrunkIpAccessControlList extends SpeakeasyBase {
  @Metadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @Metadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @Metadata({ data: "json, name=date_updated" })
  dateUpdated?: Date;

  @Metadata({ data: "json, name=friendly_name" })
  friendlyName?: string;

  @Metadata({ data: "json, name=sid" })
  sid?: string;

  @Metadata({ data: "json, name=trunk_sid" })
  trunkSid?: string;

  @Metadata({ data: "json, name=url" })
  url?: string;
}